fix: Linting issues

+++ renderer/structs/info/sync/fenceCreate.h
@@ -9,8 +9,9 @@ public:
   FenceCreate(bool signaled = false)
       : VkFenceCreateInfo{.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO,
                           .pNext = nullptr,
-                          .flags =
-                              signaled ? VK_FENCE_CREATE_SIGNALED_BIT : 0u} {}
+                          .flags = signaled ? static_cast<uint32_t>(
+                                                  VK_FENCE_CREATE_SIGNALED_BIT)
+                                            : 0u} {}
 };
 } // namespace info
 } // namespace vk
